Refactor cert error handling to act on single errors
With QtNetwork, we can get multiple errors at once (reproduced e.g. by
using the Superfish certificate on badssl.com). With QtWebEngine, we can
only get one error.

Things like "is this error coming from a third-party resource" are hard
to reason about if there is a list of errors, so change
CertificateErrorWrapper so that it's exposed as one single error to the
rest of the code.