Make notifications mostly work.
`QVariant.Type` has moved to `QMetaType.Type`[1][] and QMeta.Type
doesn't work with int().

`QImage(':/icons/qutebrowser-64x64.png')` yields and empty QImage. This
is not fixed but things don't crash because of it anymore. For instance
the "Title & Body" test on https://web-push-book.gauntface.com/demos/notification-examples/

`QtWebEngineCore.QWebEnginePage.Feature` doesn't work with int(), so add
it to the maps twice.

`for s in scripts` was from a previous hack
[1]: https://www.qt.io/blog/whats-new-in-qmetatype-qvariant