sql: Make Query.bound_values() return a dict again
SqlQuery.boundValues() changed in Qt6 from returning a map of
placeholders -> values to just providing a list of values.

We can either:

1. follow that change and always return a list
    * this has the effect of making the return value have more items if
      the caller uses a placeholder more than once in a query, like
      histcategory does
2. maintain the current return type

Here I have chosen to do (2). Although (1) is an option since it looks
like no caller actually cares about the contents of the response at this
point (apart from tests), just that it is Truthy/Falsy and the right
length (if Truthy). And callers should know how many times they re-used
placeholders so should be able to adjust their length comparison themselves.

There is no API for enumerating placeholder labels anymore, if we
want to maintain our API we must either 1) store the placeholders so we
can look them up based on label later, 2) guess what they are (eg assume
the caller always used sequential integers starting at 0), or 3) always
return the dict keys as the positional indexes instead of labels (and
return a larger dict if the caller used single placeholders multiple
times).

I've chosen to do 3 which should have the same result as the 5.15
implementation, at the cost of some more list allocations.
So now we store the placeholders so we can query for them directly
later.
And the existing tests should all pass, if you can get them to run.

This approach works on both 5.15 and 6.2, so no version checks
necessary.